New argparse (#55)
* added commandline inputs to separate file. Reach_parallel doesn't work yet because of waterbodies addon.

* working reach_parallel file but resevoir t/f not working just yet

* notebook for reading netcdf files from WRF

* removed netcdf file for now

* blacked files

* removed import argparse
