+++ src/python_framework/nhd_network_traversal.py
@@ -8,77 +8,12 @@ import recursive_print
 import time
 import os
 import argparse
-
-
-def _handle_args():
-    parser = argparse.ArgumentParser(
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter
-    )
-    parser.add_argument(
-        "--debuglevel",
-        help="Set the debuglevel",
-        dest="debuglevel",
-        choices=["0", "1", "2", "3"],
-        default=0,
-    )
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        help="Verbose output (leave blank for quiet output)",
-        dest="verbose",
-        action="store_true",
-    )
-    parser.add_argument(
-        "-t",
-        "--showtiming",
-        help="Set the showtiming (leave blank for no timing information)",
-        dest="showtiming",
-        action="store_true",
-    )
-    parser.add_argument(
-        "-n",
-        "--supernetwork",
-        help="Choose from among the pre-programmed supernetworks (Pocono_TEST1, Pocono_TEST2, LowerColorado_Conchos_FULL_RES, Brazos_LowerColorado_ge5, Brazos_LowerColorado_FULL_RES, Brazos_LowerColorado_Named_Streams, CONUS_ge5, Mainstems_CONUS, CONUS_Named_Streams, CONUS_FULL_RES_v20",
-        choices=[
-            "custom",
-            "Pocono_TEST1",
-            "Pocono_TEST2",
-            "LowerColorado_Conchos_FULL_RES",
-            "Brazos_LowerColorado_ge5",
-            "Brazos_LowerColorado_FULL_RES",
-            "Brazos_LowerColorado_Named_Streams",
-            "CONUS_ge5",
-            "Mainstems_CONUS",
-            "CONUS_Named_Streams",
-            "CONUS_FULL_RES_v20",
-        ],
-        # TODO: accept multiple or a Path (argparse Action perhaps)
-        # action='append',
-        # nargs=1,
-        dest="supernetwork",
-        default="Pocono_TEST1",
-    )
-
-    parser.add_argument(
-        "-f",
-        "--customnetworkfile",
-        dest="customnetworkfile",
-        help="OR... if 'custom' is chosen for the supernetwork, please enter the path of a .json file containing the supernetwork information. See test/input/json/CustomInput.json for an example.",
-    )
-
-    args = parser.parse_args()
-
-    if args.supernetwork == "custom" and not args.customnetworkfile:
-        parser.error(
-            r"If 'custom' is selected for the supernetwork, you must enter a path to a supernetwork-describing .json file"
-        )
-
-    return args
+import nwm_network_commandline as cmd
 
 
 def main():
 
-    args = _handle_args()
+    args = cmd._handle_args()
 
     # find the path of the test scripts, several levels above the script path
     root = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
